refactor event page
