+++ components/layout/MainLayout.tsx
@@ -1,6 +1,7 @@
+import React from "react";
+
 import Header from "components/Header";
 import Footer from "components/Footer";
-import React from "react";
 
 export default function MainLayout({
   children,
