+++ lib/types.ts
@@ -4,3 +4,10 @@ export interface ITestimony {
   school: string;
   title?: string;
 }
+
+export type IHMCEvent = {
+  id: number;
+  name: string;
+  date: string;
+  image: string;
+};
