+++ pages/curriculum.tsx
@@ -1,4 +1,4 @@
-import MainLayout from "components/layout/MainLayout";
+import MainLayout from "components/Layout/MainLayout";
 
 export default function CurriculumPage() {
   return (
