+++ pages/event/index.tsx
@@ -1,56 +1,17 @@
-import MainLayout from "components/layout/MainLayout";
-import { formatDate } from "lib/date";
 import { GetServerSidePropsContext } from "next";
-import Link from "next/link";
 
-function EventItem({ locale, event }: { locale: string; event: HMCEvent }) {
-  return (
-    <div data-id={event.id} className="p-4 border-b">
-      <div className="flex">
-        <div
-          style={{ width: 160, height: 86 }}
-          className="mr-4 bg-gray-200 rounded"
-        ></div>
-        <div>
-          <p className="text-yellow-700">{formatDate(event.date, locale)}</p>
-          <h2 className="font-bold text-xl">
-            <Link href={`/event/${event.id}`}>
-              <a>{event.name}</a>
-            </Link>
-          </h2>
-          <p className="mb-4">Online</p>
-        </div>
-      </div>
-    </div>
-  );
-}
+import MainLayout from "components/Layout/MainLayout";
+import EventFilter from "./EventFilter";
+import EventItem from "./EventItem";
 
-function EventFilter() {
-  return (
-    <div className="event-filter flex">
-      <div className="mr-4">
-        <select>
-          <option>In Person</option>
-          <option>Online</option>
-        </select>
-      </div>
-      <div>
-        <select>
-          <option>Tech</option>
-          <option>Design</option>
-          <option>Entrepreneruship</option>
-        </select>
-      </div>
-    </div>
-  );
-}
+import { IHMCEvent } from "lib/types";
 
 export default function EventListPage({
   locale,
   events,
 }: {
   locale: string;
-  events: HMCEvent[];
+  events: IHMCEvent[];
 }) {
   return (
     <MainLayout>
@@ -67,17 +28,10 @@ export default function EventListPage({
   );
 }
 
-type HMCEvent = {
-  id: number;
-  name: string;
-  date: string;
-  image: string;
-};
-
 export async function getServerSideProps(
   ctx: GetServerSidePropsContext & { locale: string }
 ) {
-  let events: HMCEvent[] = [];
+  let events: IHMCEvent[] = [];
   try {
     events = await fetch(
       process.env.NEXT_PUBLIC_API_BASE_URL + "/api/v1/events"
