+++ pages/index.tsx
@@ -1,6 +1,6 @@
 import Head from "next/head";
 
-import MainLayout from "components/layout/MainLayout";
+import MainLayout from "components/Layout/MainLayout";
 import HeroSection from "components/HeroSection";
 import AboutUsSection from "components/AboutUsSection";
 import ProgramsSection from "components/ProgramsSection";
